step03 done

+++ src/main/java/org/homework/TodoService.java
@@ -1,5 +1,6 @@
 package org.homework;
 
+import java.nio.file.attribute.UserPrincipalNotFoundException;
 import java.util.IllformedLocaleException;
 import java.util.List;
 import java.util.Map;
@@ -13,9 +14,8 @@ public class TodoService {
         this.todoRepository = new TodoRepository();
     }
 
-    public Todo findById(int id) {
-        return todoRepository.findById(id)
-                .orElseThrow(RuntimeException::new);
+    public Optional<Todo> findById(int id){
+        return todoRepository.findById(id);
     }
     public int save(Todo todo) {
         return todoRepository.save(todo);
@@ -24,8 +24,12 @@ public class TodoService {
     public boolean existsById(int id) {
         return todoRepository.existsById(id);
     }
-    public void delete(int id) {
-        todoRepository.deleteById(id);
+    public boolean delete(int id) {
+        if (todoRepository.existsById(id)) {
+            todoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
     public void update(Todo todo) {
         todoRepository.update(todo);
